# api/openapi.yaml
openapi: 3.1.0
info:
  title: shop-inventory-app API
  version: 0.2.0
  description: |
    Multi-channel inventory back-office. A single stock pool is sold through
    several channels (store, Shopee, LINE OA).

    Stock is tracked as on_hand and reserved, where available = on_hand -
    reserved. An order reserves every line or none of them, inside one
    transaction that locks the balance rows, so two channels can never sell
    the same unit.

    Every endpoint under /api except /api/auth/login and /api/auth/logout
    needs the session cookie set by login. Owner-only endpoints answer 403 for
    staff.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - session: []
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Sign in and receive the session cookie
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Signed in. Sets the HttpOnly, SameSite=Lax `sid` cookie for 7 days.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: Wrong email or password (code invalid_credentials). The same answer for unknown emails.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422": { $ref: "#/components/responses/Validation" }
...
  /api/auth/me:
    get:
      tags: [auth]
      summary: The signed-in user
      operationId: me
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/orders:
    get:
      tags: [orders]
      summary: Orders, newest first
      operationId: listOrders
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/OrderStatus" }
        - name: q
          in: query
          description: Matches order number or channel reference.
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: A page of orders.
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/OrderSummary" }
                  total: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [orders]
      summary: Create an order and reserve its stock, all or nothing
      description: |
        Locks the balance rows of every line in product id order, checks each
        line against available stock, then reserves them all in one
        transaction. If any line is short nothing is reserved, the attempt is
        written to the audit log, and the response lists every short line.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                channel: { $ref: "#/components/schemas/Channel" }
                external_ref: { type: string, maxLength: 100, description: Order id on the channel; unique per channel. }
                note: { type: string, maxLength: 500 }
                items:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items:
                    type: object
                    required: [product_id, qty]
                    properties:
                      product_id: { type: integer, format: int64 }
                      qty: { type: integer, minimum: 1, maximum: 100000 }
      responses:
        "201":
          description: Created with status reserved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: |
            insufficient_stock with details.items listing every short line, or
            conflict when the channel reference was already recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              examples:
                short:
                  value:
                    error:
                      code: insufficient_stock
                      message: stock changed and some items no longer have enough available
                      details:
                        items:
                          - { product_id: 2, sku: SKU-0002, name: กางเกงยีนส์ขายาว 32, requested: 3, available: 2 }
        "422": { $ref: "#/components/responses/Validation" }
  /api/orders/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [orders]
      summary: One order with its lines
      operationId: getOrder
      responses:
        "200":
          description: Order.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/orders/{id}/pack:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [orders]
      summary: reserved to packed; stock is unchanged
      operationId: packOrder
      responses:
        "200": { $ref: "#/components/responses/OrderOk" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/InvalidState" }
  /api/orders/{id}/ship:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [orders]
      summary: packed to shipped; on_hand and reserved both drop (SHIP movements)
      operationId: shipOrder
      responses:
        "200": { $ref: "#/components/responses/OrderOk" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/InvalidState" }
  /api/orders/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [orders]
      summary: reserved or packed to canceled; reserved units are released (RELEASE movements)
      operationId: cancelOrder
      responses:
        "200": { $ref: "#/components/responses/OrderOk" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/InvalidState" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: sid
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  responses:
    Unauthorized:
      description: No valid session (code unauthorized).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    NotFound:
      description: No such resource (code not_found).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Validation:
      description: The request body failed validation (code validation_failed).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    InvalidState:
      description: The order's current status does not allow this action (code invalid_state).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    OrderOk:
      description: The order after the transition.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Order" }
  schemas:
...
    Role:
      type: string
      enum: [owner, staff]
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
...
    Money:
      type: string
      pattern: '^\d{1,10}\.\d{2}$'
      description: Baht as a decimal string, never a float.
      examples: ["600.00"]
...
    Channel:
      type: string
      enum: [store, shopee, line]
      default: store
    OrderStatus:
      type: string
      enum: [reserved, packed, shipped, canceled]
    OrderItem:
      type: object
      required: [product_id, sku, name, qty, unit_price]
      properties:
        product_id: { type: integer, format: int64 }
        sku: { type: string }
        name: { type: string }
        qty: { type: integer }
        unit_price: { $ref: "#/components/schemas/Money" }
    OrderSummary:
      type: object
      required: [id, order_no, channel, status, total, item_count, created_at]
      properties:
        id: { type: integer, format: int64 }
        order_no: { type: string, examples: [ORD-2026-00012] }
        channel: { $ref: "#/components/schemas/Channel" }
        external_ref: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/OrderStatus" }
        total: { $ref: "#/components/schemas/Money" }
        item_count: { type: integer, description: Sum of line quantities. }
        created_by_name: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
    Order:
      type: object
      required: [id, order_no, channel, status, total, items, created_at, updated_at]
      properties:
        id: { type: integer, format: int64 }
        order_no: { type: string }
        channel: { $ref: "#/components/schemas/Channel" }
        external_ref: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/OrderStatus" }
        total: { $ref: "#/components/schemas/Money" }
        note: { type: [string, "null"] }
        created_by_name: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        packed_at: { type: [string, "null"], format: date-time }
        shipped_at: { type: [string, "null"], format: date-time }
        canceled_at: { type: [string, "null"], format: date-time }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
...
    ErrorResponse:
      type: object
      required: [error]
      description: The envelope every failing endpoint returns.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code. Clients branch on this, not on message.
              enum:
                - bad_request
                - validation_failed
                - not_found
                - internal_error
                - service_unavailable
                - unauthorized
                - forbidden
                - conflict
                - invalid_credentials
                - insufficient_stock
                - invalid_state
            message:
              type: string
              description: Human-readable explanation. Not intended for programmatic use.
            request_id:
              type: string
              description: Correlates this response with the server logs.
            fields:
              type: array
              description: Present only for validation_failed.
              items:
                $ref: "#/components/schemas/FieldError"
            details:
              description: Code-specific data, such as the short lines of insufficient_stock.
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string